fix(landing): fix LandingSection.Heading crashing under Server Components
Same root cause as DashboardSection: LandingSection.Heading is a static property attached to the exported function after the fact, so React Server Components lost it whenever <LandingSection.Heading> rendered from a Server Component. This crashed the home page, since neither the hero nor the CTA section is a client component.

Give LandingSectionHeading its own 'use client' directive and import it by name in hero.tsx and cta.tsx instead of through the LandingSection.Heading property. Break the section.tsx <-> compounds/heading/heading.tsx circular import by moving the context and hook into a sibling section-context.ts.
