+++ src/module/landing/ui/sections/cta/cta.tsx
@@ -1,3 +1,4 @@
+import { LandingSectionHeading } from '@/landing/ui/section/compounds/heading/heading';
 import { LandingSection } from '@/landing/ui/section/section';
 import { LinkButton } from '@/ui/link-button/link-button';
 
@@ -5,12 +6,12 @@ export function CtaSection() {
   return (
     <LandingSection className="flex h-[50vh] flex-col items-center justify-center gap-10 text-center">
       <div className="flex flex-col items-center gap-3">
-        <LandingSection.Heading
+        <LandingSectionHeading
           className="text-xl font-medium md:text-2xl"
           headingLevel="h2"
         >
           Keep every car, every service, and every deadline in one place.
-        </LandingSection.Heading>
+        </LandingSectionHeading>
         <p className="text-alpha-grey-700 text-sm md:text-base">
           Start for free.
         </p>
